moved HTML-adjacent code into Sienar.Html namespace
