+++ playground/TestProject.Client.Wasm/TestProjectClientPlugin.cs
@@ -1,5 +1,6 @@
 ﻿using Sienar.Configuration;
 using Sienar.Extensions;
+using Sienar.Html;
 using Sienar.Infrastructure;
 using Sienar.Layouts;
 using Sienar.Plugins;
