+++ src/Sienar.Architecture.Wasm/Infrastructure/WasmApplicationAdapter.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Sienar.Html;
 using Sienar.Plugins;
 
 namespace Sienar.Infrastructure;
