+++ src/Sienar.Architecture.Web/Infrastructure/WebApplicationAdapter.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Sienar.Html;
 using Sienar.Plugins;
 
 namespace Sienar.Infrastructure;
