+++ src/Sienar.Architecture.Web/Ui/SienarApp.razor
@@ -1,10 +1,12 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Sienar.Extensions
+@using Sienar.Html
 @using static Microsoft.AspNetCore.Components.Web.RenderMode
-@using Sienar.Infrastructure
 
+@* ReSharper disable CollectionNeverUpdated.Local *@
 @inject StyleProvider StyleProvider
 @inject ScriptProvider ScriptProvider
+@* ReSharper restore CollectionNeverUpdated.Local *@
 
 <!DOCTYPE html>
 <html lang="en">
