+++ src/Sienar.Ui/Plugins/MudBlazorPlugin.cs
@@ -1,5 +1,6 @@
 ﻿using MudBlazor.Services;
 using Sienar.Configuration;
+using Sienar.Html;
 using Sienar.Infrastructure;
 
 namespace Sienar.Plugins;
