+++ src/Sienar.Utils/Extensions/EnumExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Reflection;
+using Sienar.Html;
 using Sienar.Infrastructure;
 
 namespace Sienar.Extensions;
