+++ src/Sienar.Utils/Ui/SienarHead.razor
@@ -1,5 +1,4 @@
-@using Sienar.Extensions
-@using Sienar.Infrastructure
+@using Sienar.Html
 @* ReSharper disable CollectionNeverUpdated.Local *@
 @inject StyleProvider StyleProvider
 @inject ScriptProvider ScriptProvider
